client/requirements.txt
altgraph==0.17.4
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
importlib_metadata==7.1.0
packaging==24.0
pkg_resources==0.0.0
py-machineid==0.5.1
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
Pysher==1.0.8
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
websocket-client==1.8.0
zipp==3.18.1
